some clean-up in calendar code and doc

+++ lib/src/calendar.h
@@ -110,7 +110,7 @@ int gretl_alt_strfdate (char *s, int slen, int julian,
 			guint32 ed);
 
 int gretl_strftime (char *s, int slen, const char *format,
-		    gint64 t, int off_secs);
+		    gint64 t, double off_secs);
 
 char *gretl_strptime (const char *s, const char *format,
 		      double *dt);
